Refactor, move header management functions to their own file
