+++ internal/messaging/pubsub.go
@@ -2,15 +2,12 @@ package messaging
 
 import (
 	"fmt"
-	"strings"
 
 	"google.golang.org/protobuf/proto"
 
 	"go.eloylp.dev/goomerang/internal/messaging/protocol"
 	"go.eloylp.dev/goomerang/message"
 )
-
-const userMessageHeadersPrefix = "message-"
 
 func MessageForPublish(topic string, msg *message.Message) (*message.Message, error) {
 	data, err := proto.Marshal(msg.Payload)
@@ -24,14 +21,6 @@ func MessageForPublish(topic string, msg *message.Message) (*message.Message, er
 	})
 	pubMsg.Header = packHeaders(msg.Header)
 	return pubMsg, nil
-}
-
-func packHeaders(headers message.Header) message.Header {
-	packedHeaders := message.Header{}
-	for k, v := range headers {
-		packedHeaders.Set(fmt.Sprintf("%s%s", userMessageHeadersPrefix, k), v)
-	}
-	return packedHeaders
 }
 
 func MessageFromPublish(registry message.Registry, msg *message.Message) (*message.Message, error) {
@@ -49,15 +38,4 @@ func MessageFromPublish(registry message.Registry, msg *message.Message) (*messa
 	pubMsg := message.New().SetPayload(clientMsg)
 	pubMsg.Header = unpackHeaders(msg.Header)
 	return pubMsg, nil
-}
-
-func unpackHeaders(headers message.Header) message.Header {
-	unpackedHeaders := message.Header{}
-	for k, v := range headers {
-		if strings.HasPrefix(k, userMessageHeadersPrefix) {
-			key := strings.TrimPrefix(k, userMessageHeadersPrefix)
-			unpackedHeaders.Set(key, v)
-		}
-	}
-	return unpackedHeaders
 }
